Move dataset stuff to slice
